% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{api_plus_one}
\alias{api_plus_one}
\title{Demonstrating roxygen documentation for \code{api_plus_one()} C++ function}
\usage{
api_plus_one(x = numeric())
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector where 1 has been added to each element of x.
}
\description{
You could also use roxygen comments directly in the .cpp source (see vignette Rcpp-attributes).
}
