% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{dog}
\alias{dog}
\title{Demonstration of exposed C++ class}
\usage{
dog(name)
}
\arguments{
\item{name}{name of the dog}
}
\description{
Generated with \code{Rcpp::exposeClass()} in \code{setup_class.R}. A dog with specific
name, ears, size and the ability to bark.
}
\section{Fields}{

\describe{
\item{\code{name}}{name of the dog}
}}

\section{Methods}{

\describe{
\item{\code{bark()}}{The dog barks.}
\item{\code{setDogSize(character())}}{Set the dog's size.}
\item{\code{setEars(character())}}{Set the dog's ears.}
}
}

\examples{
singne <- dog("Singne")
singne$setDogSize("medium")
singne$setEars("floppy")
singne$bark()
}
\seealso{
\code{setup_class.R}, \code{src/api_class.h} and \code{src/api_class.cpp}
}
