% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{RcppScratch}
\alias{RcppScratch}
\alias{RcppScratch-package}
\title{Some learnings for wrapping C++ code with Rcpp}
\description{
My philosophy is to be as generic in C++ code as possible without mixing R(cpp)
specifics into the source. This can be achieved by writing an api header and
source which then can be included in another cpp file wrapping the api
functionality (i.e. preparing the input from R, casting the Rcpp datatpyes to
generic ones as well as the return values)...
}
\section{Intention}{

The following topics are explored in this package:
\itemize{
\item How to expose C++ functions (in particular \code{src/api.h}, \code{src/api.cpp} and
\code{src/hello_api.cpp}). Each function explores a point of interest. I.e. what
happens if we call \code{exit()}, can we use \code{stdout}, etc.
\item How to expose a C++ class (see \code{src/api_class.h}, \code{src/api_class.cpp} and
\code{setup_class.R}).
\item Experiments with error handling in \code{error_handling.R} (\code{cpp_from_R.R}) and
\item Some other minor questions I had in \code{cpp_from_R.R} (which also illustrates
the use of \code{Rcpp::cppFunction()} and \code{Rcpp::sourceCpp()}).
\item Chapter 16 of \emph{Extending R} (which is very nice) is summarized in the
vignette.
}
}

\section{On using \code{roxygen2}}{

\strong{Note:} Not all functions are documented. Just serves as an illustration.

When using \code{roxygen2} you can delete the NAMESPACE file, add \code{Encoding: UTF-8}
to the \code{DESCRIPTION} file (and \code{Roxygen: list(markdown = TRUE)}) as well as
the autogenerated directives by \code{Rcpp} (see first lines of \code{R/roxygen.R}).
}

\seealso{
Extending R Chapter 16 and Rcpp vignettes.
}
